requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.4.0
asyncio==3.4.3
attrs==24.2.0
Authlib==1.3.1
billiard==4.2.1
blinker==1.9.0
cachelib==0.13.0
cargo==0.3
celery==5.4.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
cryptography==43.0.0
dataclasses-json==0.6.7
distro==1.9.0
et_xmlfile==2.0.0
fastembed==0.6.0
filelock==3.18.0
Flask==3.1.0
Flask-Session==0.8.0
flatbuffers==25.2.10
frozenlist==1.4.1
fsspec==2025.3.0
greenlet==3.0.3
grpcio==1.69.0
grpcio-health-checking==1.69.0
grpcio-tools==1.69.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.29.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.4.2
langchain==0.2.15
langchain-community==0.2.15
langchain-core==0.2.37
langchain-openai==0.1.23
langchain-text-splitters==0.2.2
langchain-weaviate==0.0.2
langsmith==0.1.108
loguru==0.7.3
MarkupSafe==3.0.2
marshmallow==3.22.0
mmh3==5.1.0
mpmath==1.3.0
msgspec==0.19.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
onnxruntime==1.21.0
openai==1.43.0
openpyxl==3.1.5
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==11.1.0
portalocker==2.10.1
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
py_rust_stemmers==0.1.5
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.8.0
pydantic_core==2.20.1
pypiwin32==223
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==310
PyYAML==6.0.2
qdrant-client==1.13.3
redis==5.2.1
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==74.1.0
simsimd==4.4.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sympy==1.13.3
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.21.1
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
validators==0.34.0
vine==5.1.0
wcwidth==0.2.13
weaviate==0.1.2
weaviate-client==4.11.0
Werkzeug==3.1.3
win32_setctime==1.2.0
yarl==1.9.7
zstandard==0.23.0
